+++ packages/framework/src/modules/cms/models/blocks/article-list.model.ts
@@ -3,13 +3,15 @@ import { Block, RichText } from '@/utils/models';
 export class ArticleListBlock extends Block.Block {
     title?: string;
     description?: RichText.RichText;
-    categoryId?: string;
+    categorySlug?: string;
     articleIds?: string[];
+    articlesToShow?: number;
     parent?: {
         slug: string;
     };
     labels!: {
         today: string;
         yesterday: string;
+        seeAllArticles: string;
     };
 }