+++ packages/integrations/mocked/src/modules/cms/mappers/blocks/cms.order-list.mapper.ts
@@ -99,9 +99,9 @@ const MOCK_ORDER_LIST_BLOCK_EN: CMS.Model.OrderListBlock.OrderListBlock = {
         today: 'Today',
         yesterday: 'Yesterday',
         showMore: 'Show more',
-        reorder: 'Reorder',
         clickToSelect: 'Click to select',
     },
+    reorderLabel: 'Reorder',
     detailsUrl: '/orders/{id}',
 };
 
@@ -204,9 +204,9 @@ const MOCK_ORDER_LIST_BLOCK_DE: CMS.Model.OrderListBlock.OrderListBlock = {
         today: 'Heute',
         yesterday: 'Gestern',
         showMore: 'Mehr anzeigen',
-        reorder: 'Erneut bestellen',
         clickToSelect: 'Klicken Sie, um auszuwählen',
     },
+    reorderLabel: 'Erneut bestellen',
     detailsUrl: '/bestellungen/{id}',
 };
 
@@ -309,19 +309,19 @@ const MOCK_ORDER_LIST_BLOCK_PL: CMS.Model.OrderListBlock.OrderListBlock = {
         today: 'Dzisiaj',
         yesterday: 'Wczoraj',
         showMore: 'Pokaż więcej',
-        reorder: 'Zamów ponownie',
         clickToSelect: 'Kliknij, aby wybrać',
     },
+    reorderLabel: 'Zamów ponownie',
     detailsUrl: '/zamowienia/{id}',
 };
 
 export const mapOrderListBlock = (locale: string): CMS.Model.OrderListBlock.OrderListBlock => {
     switch (locale) {
         case 'de':
-            return { ...MOCK_ORDER_LIST_BLOCK_DE, detailsUrl: '/bestellungen/{id}' };
+            return { ...MOCK_ORDER_LIST_BLOCK_DE };
         case 'pl':
-            return { ...MOCK_ORDER_LIST_BLOCK_PL, detailsUrl: '/zamowienia/{id}' };
+            return { ...MOCK_ORDER_LIST_BLOCK_PL };
         default:
-            return { ...MOCK_ORDER_LIST_BLOCK_EN, detailsUrl: '/orders/{id}' };
+            return { ...MOCK_ORDER_LIST_BLOCK_EN };
     }
 };