+++ packages/integrations/mocked/src/modules/cms/mappers/blocks/cms.service-list.mapper.ts
@@ -24,6 +24,8 @@ const MOCK_SERVICE_LIST_BLOCK_EN: CMS.Model.ServiceListBlock.ServiceListBlock = 
         paymentPeriod: {
             MONTHLY: 'mo',
             YEARLY: 'ye',
+            WEEKLY: 'we',
+            ONE_TIME: 'one-time',
         },
     },
     pagination: {
@@ -106,8 +108,10 @@ const MOCK_SERVICE_LIST_BLOCK_DE: CMS.Model.ServiceListBlock.ServiceListBlock = 
             EXPIRED: 'Abgelaufen',
         },
         paymentPeriod: {
-            MONTHLY: 'mo',
-            YEARLY: 'ye',
+            MONTHLY: 'monatlich',
+            YEARLY: 'jährlich',
+            WEEKLY: 'wochenweise',
+            ONE_TIME: 'einmalig',
         },
     },
     pagination: {
@@ -192,6 +196,8 @@ const MOCK_SERVICE_LIST_BLOCK_PL: CMS.Model.ServiceListBlock.ServiceListBlock = 
         paymentPeriod: {
             MONTHLY: 'miesięczny',
             YEARLY: 'roczny',
+            WEEKLY: 'tygodniowy',
+            ONE_TIME: 'jednorazowy',
         },
     },
     pagination: {