+++ packages/integrations/algolia/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @o2s/integrations.algolia
 
+## 1.1.0
+
+### Minor Changes
+
+- 84510e2: adjusted algolia article search
+
+## 1.0.0
+
+### Major Changes
+
+- 0e0c816: Official stable release
+
+### Patch Changes
+
+- Updated dependencies [0e0c816]
+    - @o2s/framework@1.0.0
+    - @o2s/utils.logger@1.0.0
+
 ## 0.6.0
 
 ### Minor Changes